Changing func_get_args
Due to PHP versions prior to 5.3.0, I changed func_get_args() to be set to a parameter. Following is the note from http://php.net/manual/en/function.func-get-args.php "Because this function depends on the current scope to determine parameter details, it cannot be used as a function parameter in versions prior to 5.3.0. If this value must be passed, the results should be assigned to a variable, and that variable should be passed."